Moved `impl Request for CallRegistryGetChangesSinceRequest` from agent to the test where it is used. This is maybe unconventional, but it allows us to mark the fetch_large_record_test_canister library as testonly, which is considered "highly desirable".
