+++ rs/nervous_system/agent/Cargo.toml
@@ -21,7 +21,6 @@ ic-node-rewards-canister-api = { path = "../../node_rewards/canister/api" }
 cycles-minting-canister = { path = "../../nns/cmc" }
 ic-nns-constants = { path = "../../nns/constants" }
 ic-registry-canister-api = { path = "../../registry/canister/api" }
-ic-registry-fetch-large-record-test-canister = { path = "../../registry/fetch_large_record_test_canister" }
 ic-sns-wasm = { path = "../../nns/sns-wasm" }
 ic-sns-governance-api = { path = "../../sns/governance/api" }
 pocket-ic = { path = "../../../packages/pocket-ic" }
